requirements.txt
numpy==1.24.4
pygame==2.5.2
streamlit==1.29.0
python-dotenv==1.0.1
gtts==2.5.1
sounddevice==0.4.7
deepgram-sdk==2.12.0 #ensure they are installed in the system
groq==0.9.0 #ensure they are installed in the system
flask==2.2.5
